Add files via upload
Completed file of the texas correspondence visualization. Credit to Dr. Graham and Jeff Blackadar for getting me on the right track with uploading the file. This process was fairly easy once I got started and I can easily see how it might be useful. These visualizations are ultimately at my discretion how they may appear. This fact is always something to keep in mind when looking at these visualizations but they are still helpful especially when the research data is also available in its raw form. This one would need more cleaning before it could be shown as being a useful visual but now I know some of the basics I would be able to do this for other data.